Rename qspi_amoled to qspi_dbi
