+++ tests/components/qspi_dbi/test.esp32-s3-idf.yaml
